[XLA:GPU] Split GPU module globals files
